+++ galley/pkg/config/analysis/local/analyze.go
@@ -37,20 +37,20 @@ const domainSuffix = "svc.local"
 
 // SourceAnalyzer handles local analysis of k8s and file based event sources
 type SourceAnalyzer struct {
-	m                  *schema.Metadata
-	sources            []event.Source
-	analyzer           analysis.Analyzer
-	transformProviders transformer.Providers
+	m                    *schema.Metadata
+	sources              []event.Source
+	analyzer             analysis.Analyzer
+	transformerProviders transformer.Providers
 }
 
 // NewSourceAnalyzer creates a new SourceAnalyzer with no sources. Use the Add*Source methods to add sources in ascending precedence order,
 // then execute Analyze to perform the analysis
 func NewSourceAnalyzer(m *schema.Metadata, analyzer analysis.Analyzer) *SourceAnalyzer {
 	return &SourceAnalyzer{
-		m:                  m,
-		sources:            make([]event.Source, 0),
-		analyzer:           analyzer,
-		transformProviders: transforms.Providers(m),
+		m:                    m,
+		sources:              make([]event.Source, 0),
+		analyzer:             analyzer,
+		transformerProviders: transforms.Providers(m),
 	}
 }
 
@@ -66,7 +66,7 @@ func (sa *SourceAnalyzer) Analyze(cancel chan struct{}) (diag.Messages, error) {
 
 	updater := &snapshotter.InMemoryStatusUpdater{}
 	distributor := snapshotter.NewAnalyzingDistributor(updater, sa.analyzer, snapshotter.NewInMemoryDistributor())
-	rt, err := processor.Initialize(sa.m, domainSuffix, event.CombineSources(src, meshsrc), sa.transformProviders, distributor)
+	rt, err := processor.Initialize(sa.m, domainSuffix, event.CombineSources(src, meshsrc), sa.transformerProviders, distributor)
 	if err != nil {
 		return nil, err
 	}
