+++ galley/pkg/config/processor/transforms/serviceentry/create_test.go
@@ -769,7 +769,7 @@ func newHandler() (*processing.Runtime, *fixtures.Source, *snapshotter.InMemoryD
 		DomainSuffix: domain,
 		Source:       event.CombineSources(src, meshSrc),
 		ProcessorProvider: func(o processing.ProcessorOptions) event.Processor {
-			xforms := serviceentry.GetProviders().Create(o)
+			xforms := serviceentry.GetProviders().ToTransformers(o)
 			xforms[0].DispatchFor(metadata.IstioNetworkingV1Alpha3SyntheticServiceentries, a)
 			settings := []snapshotter.SnapshotOptions{
 				{
