+++ galley/pkg/config/processor/transforms/serviceentry/handler_bench_test.go
@@ -307,5 +307,5 @@ func newBenchHandler() event.Transformer {
 		DomainSuffix: domain,
 		MeshConfig:   meshcfg.Default(),
 	}
-	return serviceentry.GetProviders().Create(o)[0]
+	return serviceentry.GetProviders().ToTransformers(o)[0]
 }
